Towards the Grand Canonical Ensemble
The SGF container appends the extra terms at the end of the
kinetic term array. The ProbabilitiesBase detects those
terms and indirectly determines the type of the ensemble.

I don't like this design but the SGFContainer and the
ProbabilitiesBase class should be the same class anyway.
I should fix that later.