Rewrite of the TSum class
Up to now, each node of the tree corresponds to a single operator and contains a number which is the total sum of this operator and the operators that are its children. Then when I chose I perform a 3 way choice: pick among the node or the left or right child. Let W be the value of a node and W-WR-WL is the matrix element of the node.
The disadvantage of this design is that the matrix element of the top nodes (that have bigger sums) is not as accurate. This is because in the top nodes WR+WL is pretty big relative to the matrix element.
This means that by design and even after a fresh rebuild, the matrix elements of the top nodes have less than double precision!

Now I decided to change this completely. I introduced a balanced tree structure with twice as many nodes as before. The deepest nodes correspond to the operators that hold the correct matrix element, whereas their ancestors contain all partial sums. This means that W=WR+WL. The choice that I have no make now is if I should go right or left.

The second design results in an update speed which is LogN+1, where N is the number of terms. The previous design is LogN. The difference will obviously vanish for large systems. For L=4 with the spin one model LogN=8 and the slow down is about 12.8%. But I suspect that I can reduce the precision of the matrix elements to double which will give a speed up of more than 30%.