Fixed a bug introduced when I rewrote TSum
This bug was introduced with commit 37bcd4a92cc8fb2935ba9b7022b8251213d73465.
In this class N terms are stored in a tree with 2P-1 where P=2^d is the smallest power of 2 which is greater than N. The partial sums are stored in the first P-1 terms and matrix elements of the terms are in the last P terms. I failed to realize that the offset was P-1 and I thought it was P.