Major simplification in the TSum
In this commit I managed to rewrite the TSum class using the requirement
that no floating point errors should be possible. This is possible
because everytime a matrix element changes at the leaf of the tree
the tree is parsed level by level from the leafs to the root and the
sums are set to be exactly the sums of the children. This is more
reliable than propagating the difference from each leaf to the node.

Because of this simplification we no longer need a Crop function to cut
small numbers, because they don't exist. Unfortunately we need to keep
the ProbabilityPeriod because the energies still need to keep updated
incrementally.