Simplified the adjacency list
I always wanted to do that. The fast_updates has been modified in such a way that the adjacency list element don't need to calculate matrix elements anymore. Therefore the adjacency list of term_a only holds those the pointer to the terms that will be affected by the insertion or removal of that term.