Better type choices
In this commit I basically eliminate the need to make the Matrix
elements long doubles which results in a significan performance boost.
The key is to only set as long doubles the quantities that are being
accumulated, such as the energies or the normalization of tsum. In tsum
the floating point errors have been eliminated and by design the tree is
in a consistent state.
The energies however are being accumulated which results in floating
point errors after many updates so right now a rebuild required. However
if I use more intelligent accmulators in the future I can eliminate the
need for rebuilding completely.
Now using long double or double for the MatrixElement should not make
any difference in the results (digit by digit).