Refactor Measurements
I removed a significant performance bug. Before, if you tried to
measure some operator that is a linear combination of many products of
creation annihilation operators, the updates are of the order of the
number of operators in the expression (think of structure factors for
example). In fact this is why removed the automatic measurements that
you put in (don't worry, they are somewhere in the repository). It turns
out that I was just being stupid. The analogy is to say that I was doing
a huge matrix multiplication every after update, whereas I only needed
to do it when I flush the bins.