Light thermalizations
In this important commit, I have eliminated the measurement
code from thermalizations. Namely the List of BrokenLines
is not updated during thermalization but only on measurements.
This is done by making the object that contains the list of
broken lines instance of a derived class of UpdatableObject,
which is a class with the property that it's objects
are automatically updated after each operator insertion/removal.