MPI version of the code
The variables that I had to reduce are:

1) The counters for the number of updates/directed updates and the total
time both for thermalization and for measurements (I forgot that). I
reduce that during printing.
2) The broken line histogram (I forgot that too). I also reduce that
during printing.
3) The bins of the various measurable quantities. I reducing that at the
end of each bin.

The other modification, as you mentioned was to set the seed to
"Seed*(Rank+1)". Every mpi code is protected by ifdef statements so that
you can still compile the code serially as you used to.
The results that I get make sense, for example if I use 4 cores, the
error is half, the number of measurements/updates/directed updates are 4
times more.