Added TermCount
TermCount is a child class of UpdatableObject
which counts the difference of insertions and
removals of an operator in the right of the
GreenOperator. This is used to measure the
superfluid density from the windings.

Also I added a HamiltonianTerm.match method
which checks if an operator matches the configuration
of broken lines. This way I can debug MeasurableTerms.