Small update in RNG and fixed a crash
I changed the definition of the random number generator
now Exponential(A) return a number in [0,1[ with
distribution exp(A x) where A can be any double.

Also In GenDiagonalEnergy there was a bug:
I was never checking if the string is empty
and that resulted in some crashes at high energies.

Also I added a missing 1/G() in the Boltzmann weight.

Finally I fixed a stupid mistake in the newtime()