Fixed a bug in the diagonal energy estimator
The fix was really simple, but it broke my head. Here is the thing, in
order to calculate the integral I have to store the energies in the
operator string. This is a bit unnatural, because each energy
corresponds to a time interval, but each operator holds a time index,
which is in the boundary of a time interval. Obviously you have to be
careful about how to handle this. For example in my convention for left
motion I have to store the right energy *after* the update, but for
right motion I need to store the energy before the update when a new
operator is created. What I was doing wrong is to store the energy
before the update in both directions. Somehow this worked in low
temperatures, probably because the length of the update is long, but
don't ask me why.

There is a better way which I did not implement yet, even though it is
probably simpler...
Instead of storing an energy for each operator, store an energy
difference. That is the energy change that the operator causes. This
makes way more sense and requires less checks during the update.

And all this is because when you have a circle and put boundaries on it
by using points, say A, B and C then there is some kind of duality
between those points and the intervals A-B, B-C and C-A.