Introduce AtomicTerm and an Example file.
AtomicTerm is an abstraction of the HamiltonianTerm class.
The parameters are the coefficient type and the IndexedProductElement.

This arrangement allows to redefine HamiltonianTerm easily for a
different type of particles.

Also I removed the Length function from Product Element since it does
nothing.

Also I introduct an example file with a 1D Hubbard model
which compiles without the need for the interface.