I merged classes in Probabilities.hh
I created a Probabilities class which contains all
data. Yes this is how it used to be and I stupidly
broke it in a set of nonsensical classes.

Anyway, the point is to set the stage for the
Grand canonical ensemble, which is hard
if the data are all over the place.