More and more rearrangements
In this commit I created the class SGFContainer
which contains all the independent data: the configuration
the alpha and beta the kinetic and potential operators
and the operator string. Objects of this class are
used to initialize the Probabilities class and the
OperatorString class. This implies that the OperatorString
is now restartable.