Initial Implementation of the Grand Canonical Ensemble
This is my first attempt for the grand canonical ensemble.
The algorithm is roughly:

if(Ensemble==Canonical or not Extra_Term_Exists or
not Diagonal_Configuration)
  choose_regular_term
else
  choose_extra_term

The extra terms are chosen uniformly randomly, the only
constraint being that their matrix elements should be
non zero. When they are chosen, obviously the
kinetic terms's trees need to be updated. Also when
another kinetic term is chosen (or the conjugate extra term)
it will affect the permissibility of those extra terms
because it can make their matrix elements zero.

In this implementation everything related to the Grand Canonical
ensemble is stored in an extrenal class called GrandCanonicalContainer.
A pointer to an object of that type is stored in the Probabilities class
If this pointer (called ensemble) is zero it means Canonical ensemble
otherwise it means Grand Canonical.
The choose() and update() methods of the Probabilities class
have been modified. If the ensemble pointer is non zero
they will try to chose and update the extra terms.

One tricky part was to figure out when we destroy an extra term.
I do this by storing the address of the extra term and compare
every term created or destroyed with it. Another way
would be to create a method which identifies extra terms
(the number of operators they have is just one),
but this would have been slightly slower.