Simplifications in Measurable.hh
I removed some dead code and also I made the class easier to use.
In particular you can now create arbitrary measurable items
outside the class. You can request the class to give you a
pointer to a buffer of any measurable operator and you can
use this in order to build arbitrary measurable functions
as long as they are members of the MeasurableFunction class.

It is for example easy to write a function which
diagonalizes the density matrix and gets the condensate.