Change in the behavior of the progress bar
I made various changes in the Simulation.hh progress bar.
I basically factored out all the ugliness into a Timer class
which keeps track of the time and the iterations
and decides if the simulation should stop or continue.

Also, the progress bar is not directed to the cerr,
which means that I can now redirect output to
files and not get a billion trash in the beginning.