Introduce a finite probability for selecting extra terms
This is the relevant to the implementation of the Grand Canonical
ensemble.
Before I was selecting an extra term in the canonical ensemble
if there was none and if we were in a diagonal configuration.
This resulting in the code hanging after a while.
To "control" this, we make sure the choice is made only sometimes
(like 5%).