Various modifications
- I added some debugging features, for example
to check if the total probabilities are zero, in which case
no choice can be made. For example if there are no particles
the total probabilities will be zero.

- I changed the formating of the output removing dummy words
such as "+/-" for the error, so that it is easier to copy-paste
the results.

- Beautified the code with astyle

- I made the GrandCanonicalContainer a bit more self contained
and hid it's internal structure.

- I created the MeasurableEigenSystem class and removed some
other useless classes.

- I fixed a bug which would make the traps non symmetric