New behavior for trapped potentials
Before any trap including the ones with spherical symmetry
were embedded in a square or cubic domain and the sites
at the corners were idle. Now I delete the sites at the corners
and I make the domain as tight as possible. This is meant to
reduce memory consumption and make the algorithm a bit
faster.