A lot of aimless refactoring and a small improvement
The small improvement is that I changed the way I evaluate
the standard deviation in Accumulator.hh. This way there
are less floating point errors.

Also I have included the potential to use the C++11 mersene
twister algorithm, but it does not work with the icc compiler
yet.

The rest is just a ton of refactorings, basically
by putting classes in separate files.

Note: The measurable class will need a lot of work still.