More aimless refactoring and some new features
The new feature is that I can print the Hamiltonian
and the operator string in a json file.
Later I will be able to read it as well.
For the moment I am just testing to see what I need
so that I don't generate the Hamiltonian, but read
it from a file.