Refactor RandomNumberGenerator.hh and new script
Default randon number generator is the local
Mersenne Twister. I confirmed that the STL one is
the same, but a bit slower for some reason.
The MPI stuff where the seed is set are moved
to the main source file. Also I refactored the WELL44497
algorithm.

There is a compile script in SHELL which has the same functionality.
as rake in order to replace rake.